Remove code duplicates.
