+++ app/src/main/java/com/nmakademija/nmaakademija/BaseActivity.java
@@ -3,6 +3,7 @@ package com.nmakademija.nmaakademija;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.view.MenuItem;
 
 import icepick.Icepick;
 
@@ -20,5 +21,16 @@ public class BaseActivity extends AppCompatActivity {
         super.onSaveInstanceState(outState);
 
         Icepick.saveInstanceState(this, outState);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case android.R.id.home:
+                this.finish();
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
     }
 }
