+++ app/src/main/java/com/nmakademija/nmaakademija/SettingsActivity.java
@@ -3,7 +3,6 @@ package com.nmakademija.nmaakademija;
 import android.os.Bundle;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
-import android.view.MenuItem;
 
 import com.nmakademija.nmaakademija.fragment.SettingsFragment;
 
@@ -18,16 +17,5 @@ public class SettingsActivity extends AppCompatActivity {
         getSupportFragmentManager().beginTransaction()
                 .replace(R.id.main_frame, new SettingsFragment())
                 .commit();
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            case android.R.id.home:
-                this.finish();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
     }
 }
